+++ doc/WhatsNew.rst
@@ -5,6 +5,7 @@ What's New
 Ver 3.2.0 (unreleased)
 ======================
 - Fixed some numpy deprecation warnings with numpy 1.19.0
+- Added an option to make a copy of existing cut in Cuts plugin
 
 Ver 3.1.0 (2020-07-20)
 ======================